requirements-test.txt
# Test requirements - install with: pip install -r requirements-test.txt

# Main application requirements
-r requirements.txt

# Testing frameworks
pytest==8.0.0
pytest-flask==1.3.0
pytest-cov==4.0.0

# Mocking and test utilities
pytest-mock==3.12.0
freezegun==1.4.0

# Code quality and coverage
coverage==7.4.0
black==24.1.1
flake8==7.0.0
isort==5.13.2

# Security testing
bandit==1.7.5
safety==3.0.1
